requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
